ksh93: import 1.0.4

+++ cde/programs/dtksh/ksh93/src/cmd/ksh93/sh/io.c
@@ -13,6 +13,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                  Martijn Dekker <martijn@inlv.org>                   *
 *          atheik <14833674+atheik@users.noreply.github.com>           *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
@@ -2297,7 +2298,7 @@ struct eval
 {
 	Sfdisc_t	disc;
 	char		**argv;
-	short		slen;
+	int		slen;
 	char		addspace;
 };
 
@@ -2348,7 +2349,7 @@ static int eval_exceptf(Sfio_t *iop,int type, void *data, Sfdisc_t *handle)
 	if(!ep->addspace)
 	{
 		/* get the length of this string */
-		ep->slen = len = strlen(cp);
+		ep->slen = len = (int)strlen(cp);
 		/* move to next string */
 		ep->argv++;
 	}
