+++ cde/programs/dtksh/ksh93/src/lib/libast/sfio/sfwr.c
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"sfhdr.h"
@@ -44,7 +45,7 @@ static ssize_t sfoutput(Sfio_t* f, char* buf, size_t n)
 			sp = buf+1;
 			if(buf[0] == 0 && buf[_Sfpage-1] == 0)
 			{	/* check byte at a time until int-aligned */
-				while(((ulong)sp)%sizeof(int))
+				while((uintptr_t)sp % sizeof(int))
 				{	if(*sp != 0)
 						goto chk_hole;
 					sp += 1;
